feat: support input and handle wrapping
This implements support for the input command. The current
implementation panics upon the enter of a non-ascii character. Wrapping
in the increment and decrement methods is added. Incrementing a cell
with 255 value will cause the cell to be set to 0. Decrementing a cell
with 0 value will cause the cell to be set to 255.